+++ projectroles/templatetags/projectroles_tags.py
@@ -230,7 +230,7 @@ def get_project_column_count(app_plugins):
             ]
         )
 
-    return 3 + max(
+    return 2 + max(
         [get_active_list_columns(app_plugin) for app_plugin in app_plugins],
         default=0,
     )