+++ projectroles/tests/test_templatetags.py
@@ -210,14 +210,20 @@ class TestCommonTemplateTags(TestTemplateTagsBase):
         """Test get_project_link()"""
         self.assertEqual(
             c_tags.get_project_link(self.project, full_title=False),
-            '<a href="/project/{}">{}</a>'.format(
-                self.project.sodar_uuid, self.project.title
+            '<a href="/project/{}" title="{}" data-toggle="tooltip" '
+            'data-placement="top">{}</a>'.format(
+                self.project.sodar_uuid,
+                self.project.description,
+                self.project.title,
             ),
         )
         self.assertEqual(
             c_tags.get_project_link(self.project, full_title=True),
-            '<a href="/project/{}">{}</a>'.format(
-                self.project.sodar_uuid, self.project.get_full_title()
+            '<a href="/project/{}" title="{}" data-toggle="tooltip" '
+            'data-placement="top">{}</a>'.format(
+                self.project.sodar_uuid,
+                self.project.description,
+                self.project.get_full_title(),
             ),
         )
         # TODO: Also test remote project link display (with icon)
@@ -484,8 +490,8 @@ class TestProjectrolesTemplateTags(TestTemplateTagsBase):
         """Test get_project_column_count()"""
         app_plugins = get_active_plugins()
 
-        self.assertEqual(tags.get_project_column_count(app_plugins), 5)
-        self.assertEqual(tags.get_project_column_count([]), 3)
+        self.assertEqual(tags.get_project_column_count(app_plugins), 4)
+        self.assertEqual(tags.get_project_column_count([]), 2)
 
     def test_get_user_role_html(self):
         """Test get_user_role_html()"""